requirements.txt
pyyaml>=6.0
matplotlib>=3.8
numpy>=1.26
requests>=2.31
openai>=1.12
